% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallete1.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{get a base list of rgb colors}
\format{An object of class \code{list} of length 1.}
\usage{
palettes
}
\value{
the list of rgb value strings
examples
pattete1()
}
\description{
get a base list of rgb colors
}
\keyword{datasets}
